Wasm support

Former-commit-id: 481dcc4c0f4b8b9c66a6f44a61db8051f54a6da0
Former-commit-id: 8679add50f635edacabcfa0fd29c09cf8b0862d4 [formerly dc69836d6aec079bf52e146119834331a39d9cb1]
Former-commit-id: ec6d9b4d444dca3cc4fdd05ba2b501f88675befb

+++ source/hip/ai/steering.d
@@ -124,8 +124,8 @@ void flocking(
 {
 
     Vector3 alignmentSteering = Vector3.zero;
-    ulong groupSize = 0;
-    for(ulong i = 0; i < groupPosition.length; i++)
+    size_t groupSize = 0;
+    for(size_t i = 0; i < groupPosition.length; i++)
     {
         if(groupPosition[i] != position && position.distance(groupPosition[i]) <= perceptionRadius)
         {
